fix: apply review comments
